+++ samples/nrf9160/coap_client/src/main.c
@@ -10,7 +10,7 @@
 #include <zephyr.h>
 #include <net/coap.h>
 #include <net/socket.h>
-#include <lte_lc.h>
+#include <modem/lte_lc.h>
 #if defined(CONFIG_LWM2M_CARRIER)
 #include <lwm2m_carrier.h>
 #endif
