+++ samples/nrf9160/http_application_update/src/main.c
@@ -7,7 +7,7 @@
 #include <drivers/gpio.h>
 #include <drivers/flash.h>
 #include <bsd.h>
-#include <lte_lc.h>
+#include <modem/lte_lc.h>
 #include <modem/at_cmd.h>
 #include <modem/at_notif.h>
 #include <net/bsdlib.h>
